perbaikan confirmasi

+++ application/views/backend/admin/listsiswa.php
@@ -43,7 +43,7 @@
 
                                         <td>
                                             <a href="<?= base_url(); ?>Admin/editSiswa/<?= $s['id']; ?>" class="btn btn-warning"><i class="fas fa-edit"></i></a>
-                                            <a href="<?= base_url(); ?>Admin/deleteSiswa/<?= $s['id']; ?>" class="btn btn-danger"><i class="fas fa-trash-alt"></i></a>
+                                            <a href="<?= base_url(); ?>Admin/deleteSiswa/<?= $s['id']; ?>" onclick="return confirm('Anda yakin akan menghapus Siswa ini?');" class="btn btn-danger"><i class="fas fa-trash-alt"></i></a>
                                         </td>
 
                                     </tr>
